requirements.txt
discord.py
python-a2s
mysql-connector-python
requests